FEI-5037: Convert `_<sub_type>.js(x)` to `.<sub_type>.ts(x)` (#4)
## Summary:
This change implements the naming convertion we want for file sub-types as described in ADR #623.  There are no tests associated with this b/c it would take a fair amount of work to standup an integration test suite to test this.  We only need these ttool for a couple handfuls of codebases.  If we notice an issue when running this on a codebase, we can tweak the logic at that point.

Issue: FEI-5037

TODO:
- [x] make this more general

## Test plan:
I tested the regex manually in node:
```
> "foo_test.js".replace(/_([^\.]+)\.(jsx?)$/, ".$1.$2")
'foo.test.js'
> "foo_testdata.jsx".replace(/_([^\.]+)\.(jsx?)$/, ".$1.$2")
'foo.testdata.jsx'
```

Author: kevinbarabash

Reviewers: somewhatabstract

Required Reviewers:

Approved By: somewhatabstract

Checks:

Pull Request URL: https://github.com/Khan/flow-to-typescript-codemod/pull/4