FEI-5075: Don't add ReactElement return type for non-component functions (#8)
## Summary:
The codemod was adding React.ReactElement as a return type to functions that take "props" as a param, but aren't functions (e.g. mapPropsToState, mapPropsToDispatch, etc.).  This PR fixes that issue.  It also, removes the code that was stripping the type annotation from the "props" param in functional components.  I forgot to remove this code in #7.

Issue: FEI-5075

## Test plan:
- yarn test

Author: kevinbarabash

Reviewers: jeresig

Required Reviewers:

Approved By: jeresig

Checks:

Pull Request URL: https://github.com/Khan/flow-to-typescript-codemod/pull/8