FEI-5154: Wrap object literals being returned from arrow functions in parens (#18)
## Summary:
We use this pattern a lot all over the place.  There was some code for wrapping the return type in parens when it's a object type.  This PR does the same but for values instead of types.

Issue: FEI-5154

## Test plan:
- yarn build
- yarn test

Author: kevinbarabash

Reviewers: jeresig

Required Reviewers:

Approved By: jeresig

Checks:

Pull Request URL: https://github.com/Khan/flow-to-typescript-codemod/pull/18